Hice funciones para archivos de high scores y cambie gameEnd para que solo la llame gameUpdate

+++ backend/scores.h
@@ -24,7 +24,7 @@
 typedef struct 
 {
 	unsigned char tag[4];
-	unsigned short int score;
+	unsigned int score;
 } score_t;
 
 /*******************************************************************************
